requirements.txt
flask
python-dotenv
requests
gspread
google-auth
yagmail